kicad-autopcb-gen/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "kicad-autopcb-gen"
version = "0.1.0"
description = "Template-driven KiCad 9 project generator (spec/prompt -> schematic + starter PCB)."
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "openai>=1.0.0",
  "kicad-sch-api>=0.1.0"
]

[project.scripts]
pcbgen = "pcbgen.cli:main"

[tool.setuptools]
packages = ["pcbgen"]
